+++ src/main/java/io/electrum/vas/model/CardPayment.java
@@ -2,6 +2,7 @@ package io.electrum.vas.model;
 
 import java.util.Objects;
 
+import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Pattern;
 
@@ -19,7 +20,14 @@ public class CardPayment extends PaymentMethod {
    private String pan = null;
    private String expiryDate = null;
    private PosInfo posInfo = null;
+   /**
+    * encrytedPin
+    *
+    * @deprecated As of v3.19.0 due to the addition of the {@link Pin Pin} model, use {@link CardPayment#pin} instead.
+    */
+   @Deprecated
    private EncryptedPin encryptedPin = null;
+   private Pin pin = null;
 
    public CardPayment() {
       setType(PaymentMethodType.CARD);
@@ -77,6 +85,15 @@ public class CardPayment extends PaymentMethod {
       this.posInfo = posInfo;
    }
 
+   /**
+    * Sets the encryptedPin field and returns the CardPayment object.
+    *
+    * @param encryptedPin
+    * @return The CardPayment object with the EncryptedPin set
+    * @deprecated As of v3.19.0 due to the addition of the {@link Pin Pin} model, use {@link CardPayment#pin(Pin pin)}
+    *             instead.
+    */
+   @Deprecated
    public CardPayment encryptedPin(EncryptedPin encryptedPin) {
       this.encryptedPin = encryptedPin;
       return this;
@@ -86,15 +103,47 @@ public class CardPayment extends PaymentMethod {
     * The encrypted PIN associated with this card in HEX format.
     *
     * @return encryptedPin
+    * @deprecated As of v3.19.0 due to the addition of the {@link Pin Pin} model, use {@link CardPayment#getPin()}
+    *             instead.
     **/
    @ApiModelProperty(value = "The encrypted pin number associated with the card in HEX format.")
    @JsonProperty("encryptedPin")
+   @Deprecated
    public EncryptedPin getEncryptedPin() {
       return encryptedPin;
    }
 
+   /**
+    * Sets the encryptedPin field.
+    *
+    * @param encryptedPin
+    * @deprecated As of v3.19.0 due to the addition of the {@link Pin Pin} model, use {@link CardPayment#setPin(Pin)}
+    *             instead.
+    */
+   @Deprecated
    public void setEncryptedPin(EncryptedPin encryptedPin) {
       this.encryptedPin = encryptedPin;
+   }
+
+   public CardPayment pin(Pin pin) {
+      this.pin = pin;
+      return this;
+   }
+
+   /**
+    * The PIN associated with this card as either a clear PIN or an encrypted PIN in HEX format.
+    *
+    * @return pin
+    */
+   @ApiModelProperty(value = "The PIN associated with this card as either a clear PIN or an encrypted PIN in HEX format.")
+   @JsonProperty("pin")
+   @Valid
+   public Pin getPin() {
+      return pin;
+   }
+
+   public void setPin(Pin pin) {
+      this.pin = pin;
    }
 
    @Override
@@ -114,7 +163,7 @@ public class CardPayment extends PaymentMethod {
       return Objects.equals(this.type, cardPayment.type) && Objects.equals(this.name, cardPayment.name)
             && Objects.equals(this.pan, cardPayment.pan) && Objects.equals(this.expiryDate, cardPayment.expiryDate)
             && Objects.equals(this.posInfo, cardPayment.posInfo)
-            && Objects.equals(this.encryptedPin, cardPayment.encryptedPin);
+            && Objects.equals(this.encryptedPin, cardPayment.encryptedPin) && Objects.equals(this.pin, cardPayment.pin);
    }
 
    @Override
@@ -129,6 +178,7 @@ public class CardPayment extends PaymentMethod {
       sb.append("    expiryDate: ").append(Utils.toIndentedString(expiryDate)).append("\n");
       sb.append("    posInfo: ").append(Utils.toIndentedString(posInfo)).append("\n");
       sb.append("    encryptedPin: ").append(Utils.toIndentedString(encryptedPin)).append("\n");
+      sb.append("    pin: ").append(Utils.toIndentedString(pin)).append("\n");
 
       sb.append("}");
       return sb.toString();